Create csv.py to separete functions about csv from word registration
